+++ kamehouse-admin/src/main/java/com/nicobrest/kamehouse/admin/controller/ScreenController.java
@@ -3,9 +3,11 @@ package com.nicobrest.kamehouse.admin.controller;
 import com.nicobrest.kamehouse.admin.model.kamehousecommand.AltTabKeyKameHouseSystemCommand;
 import com.nicobrest.kamehouse.admin.model.kamehousecommand.EnterKeyKameHouseSystemCommand;
 import com.nicobrest.kamehouse.admin.model.kamehousecommand.EscKeyKameHouseSystemCommand;
+import com.nicobrest.kamehouse.admin.model.kamehousecommand.RightKeyKameHouseSystemCommand;
 import com.nicobrest.kamehouse.admin.model.kamehousecommand.ScreenLockKameHouseSystemCommand;
 import com.nicobrest.kamehouse.admin.model.kamehousecommand.ScreenUnlockKameHouseSystemCommand;
 import com.nicobrest.kamehouse.admin.model.kamehousecommand.ScreenWakeUpKameHouseSystemCommand;
+import com.nicobrest.kamehouse.admin.model.kamehousecommand.WinTabKeyKameHouseSystemCommand;
 import com.nicobrest.kamehouse.commons.controller.AbstractSystemCommandController;
 import com.nicobrest.kamehouse.commons.model.systemcommand.SystemCommand;
 import com.nicobrest.kamehouse.commons.service.SystemCommandService;
@@ -77,5 +79,22 @@ public class ScreenController extends AbstractSystemCommandController {
   public ResponseEntity<List<SystemCommand.Output>> altTabKeyPress(
       @RequestParam(value = "tabs", required = false) Integer tabs) {
     return execKameHouseSystemCommand(new AltTabKeyKameHouseSystemCommand(tabs));
+  }
+
+  /**
+   * Sends an WIN+TAB key press on the server.
+   */
+  @PostMapping(path = "/win-tab-key-press")
+  public ResponseEntity<List<SystemCommand.Output>> winTabKeyPress() {
+    return execKameHouseSystemCommand(new WinTabKeyKameHouseSystemCommand());
+  }
+
+  /**
+   * Sends a right key press on the server.
+   */
+  @PostMapping(path = "/right-key-press")
+  public ResponseEntity<List<SystemCommand.Output>> winTabKeyPress(
+      @RequestParam(value = "keyPresses", required = false) Integer keyPresses) {
+    return execKameHouseSystemCommand(new RightKeyKameHouseSystemCommand(keyPresses));
   }
 }