+++ kamehouse-admin/src/test/java/com/nicobrest/kamehouse/admin/integration/ScreenControllerIntegrationTest.java
@@ -33,7 +33,9 @@ class ScreenControllerIntegrationTest extends AbstractControllerIntegrationTest 
       "unlockTest, /unlock",
       "escKeyTest, /esc-key-press",
       "enterKeyTest, /enter-key-press",
-      "altTabKeyTest, /alt-tab-key-press"
+      "altTabKeyTest, /alt-tab-key-press",
+      "winTabKeyTest, /win-tab-key-press",
+      "rightKeyTest, /alt-tab-key-press?keyPresses=1"
   })
   void screenControllerTest(String testName, String apiEndpoint) throws Exception {
     logger.info("Running {}", testName);